% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_reseq_basic.R
\name{bwa_mapping_plot}
\alias{bwa_mapping_plot}
\title{Plot bwa mapping stats}
\usage{
bwa_mapping_plot(mapping_stats, sample_limit, out_prefix = NULL)
}
\arguments{
\item{mapping_stats}{merged mapping stats file}

\item{sample_limit}{sample number limit to show detail inform of each sample}

\item{out_prefix}{output file prefix, default is NULL, don't output file}
}
\description{
Plot bwa mapping rate summrized by samtools
}
\details{
Mapping rate of each sample is generated by \code{samtools stats sorted.bam};

Out stats is extracted using grep and sed command:

\code{grep ^SN | cut -f 2,3 | sed -n '3p;7p;8p;9p;10p;11p;13p;15p;16p;19p;20p;23p;24p;25p;31p;32p;38p;40p'};

Stats of all samples in analysis were merged into a single table.
}
\examples{
mapping_stats <- system.file("extdata", "all_sample.mapping.xls", package = "omplotr")
# show mapping summary
bwa_mapping_plot(mapping_stats, 5)
# show detail sample information
bwa_mapping_plot(mapping_stats, 10)
}
