% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_qc.R
\name{gc_line_plot}
\alias{gc_line_plot}
\title{GC plot function}
\usage{
gc_line_plot(plot_data, output = NULL)
}
\arguments{
\item{plot_data}{gc dataframe}

\item{output}{out put prefix, default is NULL, don't output file}
}
\description{
GC plot function
}
\examples{
gc_line_plot(gc_test_data)
}
