% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quant.R
\name{om_boxplot}
\alias{om_boxplot}
\title{plot expression box, violin, density plot or merged plots of all above}
\usage{
om_boxplot(exp_data, samples, plot_type, outdir = NULL)
}
\arguments{
\item{samples}{sample <-> group map, first column is group id}

\item{plot_type}{type of plot to choose: box, violin, density, all}

\item{outdir}{out put directory, default is NULL, don't output file}

\item{plot_data}{expression matrix}
}
\description{
plot expression box, violin, density plot or merged plots of all above
}
\examples{
om_boxplot(exp_test_data, test_sample_data, 'box')
om_boxplot(exp_test_data, test_sample_data, 'violin')
om_boxplot(exp_test_data, test_sample_data, 'density')
om_boxplot(exp_test_data, test_sample_data, 'all')
}
