% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quant.R
\name{om_cluster_plot}
\alias{om_cluster_plot}
\title{plot correlation heatmap}
\usage{
om_cluster_plot(plot_data, out_prefix = NULL)
}
\arguments{
\item{plot_data}{cluster result}

\item{out_prefix}{out put prefix, default is NULL, don't output file}
}
\description{
plot correlation heatmap
}
\examples{
om_cluster_plot(cluster_test_data)
}
