% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quant.R
\name{om_correlation_plot}
\alias{om_correlation_plot}
\title{plot correlation heatmap}
\usage{
om_correlation_plot(exp_data, samples, outdir = NULL)
}
\arguments{
\item{exp_data}{expression matrix}

\item{samples}{sample <-> group map, first column is group id}

\item{outdir}{out put directory, default is NULL, don't output file}
}
\description{
plot correlation heatmap
}
\examples{
om_correlation_plot(exp_test_data, test_sample_data)
}
