% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_go.R
\name{om_enrich_bar_plot}
\alias{om_enrich_bar_plot}
\title{plot enrichment barplot}
\usage{
om_enrich_bar_plot(enrich_df, term_number = 30,
  ylab_title = "-log10(qvalue)", out_prefix = NULL)
}
\arguments{
\item{enrich_df}{GO/KEGG enrichment result table}

\item{term_number}{term number to show, default is 30}

\item{ylab_title}{text of ylab, default is "-log10(qvalue)"}

\item{out_prefix}{output plot file prefix}
}
\description{
plot enrichment barplot
}
\examples{
# kegg enrichment plot
kegg_enrich_file <- system.file("extdata", "enrichment", "example.kegg.enrichment.txt", package = "omplotr")
kegg_enrich_list <- clean_enrich_table(kegg_enrich_file)
head(kegg_enrich_list$table, 4)
om_enrich_bar_plot(kegg_enrich_list$table, ylab_title=kegg_enrich_list$title)
# go enrichment plot
go_enrich_file <- system.file("extdata", "enrichment", "example.go.enrichment.txt", package = "omplotr")
go_enrich_list <- clean_enrich_table(go_enrich_file)
om_enrich_bar_plot(go_enrich_list$table, ylab_title=go_enrich_list$title)
}
