% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_go.R
\name{om_goseq}
\alias{om_goseq}
\title{goseq analysis}
\usage{
om_goseq(diff_genes, gene_length_df, gene_go_map, out_prefix = NULL)
}
\arguments{
\item{diff_genes}{diff gene vector}

\item{gene_length_df}{gene length dataframe}

\item{gene_go_map}{gene go map file}

\item{out_prefix}{out putfile prefix, default is NULL, don't output file}
}
\description{
goseq analysis
}
\examples{
diff_genes <- go_test_data_list[['test_diff_genes']]
gene_length_df <- go_test_data_list[['test_gene_len']]
gene_go_map <- system.file("extdata", "topgo_test_data.txt", package = "omplotr")
goseq_output <- om_goseq(diff_genes, gene_length_df, gene_go_map)
head(goseq_output, 4)
}
