% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quant.R
\name{om_heatmap}
\alias{om_heatmap}
\title{plot expression heatmap}
\usage{
om_heatmap(exp_data, samples, outdir = NULL)
}
\arguments{
\item{samples}{sample <-> group map, first column is group id}

\item{outdir}{out put directory, default is NULL, don't output file}

\item{plot_data}{expression matrix}
}
\description{
plot expression heatmap
}
\examples{
om_heatmap(exp_test_data, test_sample_data)
}
