% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_go.R
\name{om_topgo}
\alias{om_topgo}
\title{topGO plot}
\usage{
om_topgo(gene_go_map, diff_genes, enrich_result_df, name = NULL,
  out_dir = NULL)
}
\arguments{
\item{gene_go_map}{gene go map file}

\item{diff_genes}{diff gene vector}

\item{enrich_result_df}{goseq analysis result}

\item{name}{out put file name, default is NULL, don't output file}

\item{out_dir}{out put directory, default is NULL, don't output file}
}
\description{
topGO plot
}
\examples{
gene_go_map <- system.file("extdata", "topgo_test_data.txt", package = "omplotr")
diff_genes <- go_test_data_list[['test_diff_genes']]
gene_length_df <- go_test_data_list[['test_gene_len']]
go_anno_df <- go_test_data_list[['test_go_anno']]
enrich_result_df <- run_goseq(diff_genes, gene_length_df, go_anno_df)
om_topgo(gene_go_map, diff_genes, enrich_result_df)
}
