% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_reseq_basic.R
\name{om_var_pie_plot}
\alias{om_var_pie_plot}
\title{Pie Plot of variant stats}
\usage{
om_var_pie_plot(sample_var_stats, color_pal = "Set2", out_prefix = NULL)
}
\arguments{
\item{sample_var_stats}{variant stats of each sample}

\item{color_pal}{color palette name, default is "Set2"}

\item{out_prefix}{output prefix, default is NULL, no output}
}
\description{
Using ggplot2 to generate pie plot of variant summary.
Variant summary data is generated using \code{om_var_pie_stats}.
}
\examples{

# pie plot data format
a_om_test_var_stats <- dplyr::filter(om_test_var_stats[[1]], variable == 'A')
head(a_om_test_var_stats)
# plot
om_var_pie_plot(a_om_test_var_stats)
}
