% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_reseq_basic.R
\name{om_var_pie_stats}
\alias{om_var_pie_stats}
\title{Transform variants summary to plot data for ggplot}
\usage{
om_var_pie_stats(stats_name, stats_dir, impact_map_file,
  stats_prefix = "all_sample", varRegion_order = NULL)
}
\arguments{
\item{stats_name}{type of summary file}

\item{stats_dir}{directory of summary file}

\item{impact_map_file}{snpeff effect impact file}

\item{stats_prefix}{prefix of summary file}

\item{varRegion_order}{plot order of variant region}
}
\description{
Transform variants summary to plot data for ggplot
}
\examples{

stats_names <- om_const_reseq_variant[['var_file_labs']][1:2]
stats_dir <- system.file("extdata", "variant_stats", package = "omplotr")
impact_map_file <- system.file("extdata", "variant_stats", "snpeff_varEffects.csv", package = "omplotr")
test_pie_stats <- lapply(stats_names, om_var_pie_stats, stats_dir=stats_dir, impact_map_file=impact_map_file)
head(test_pie_stats[[1]], 4)
}
