% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_reseq_basic.R
\name{om_var_summary_plot}
\alias{om_var_summary_plot}
\title{Variant summary plot}
\usage{
om_var_summary_plot(var_stats_df, out_prefix = NULL, width = 30,
  height = 12)
}
\arguments{
\item{var_stats_df}{variant stats dataframe}

\item{out_prefix}{output prefix, default is NULL, no output}
}
\description{
Variant portion distribution summary plot using ggplot2.
Variant summary data is generated using \code{om_var_pie_stats}.
}
\examples{

om_test_var_stats_df <- plyr::ldply(om_test_var_stats, data.frame)
om_var_summary_plot(om_test_var_stats_df)
}
