% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_quant.R
\name{om_volcano_plot}
\alias{om_volcano_plot}
\title{plot diff expression volcano plot}
\usage{
om_volcano_plot(dpa_results, compare_name, logfc = 1, qvalue = 0.05,
  outdir = NULL)
}
\arguments{
\item{dpa_results}{diff analysis results}

\item{compare_name}{compare name of diff analysis, 'ALL' for merged results}

\item{logfc}{log2foldchange cutoff for DEG analysis, default is 1}

\item{qvalue}{adjust pvalue cutoff for DEG analysis, default is 0.05}

\item{outdir}{out put directory, default is NULL, don't output file}
}
\description{
plot diff expression volcano plot
}
\examples{
om_volcano_plot(diff_test_data, 'Case_vs_Control')
om_volcano_plot(diff_test_data, 'ALL')
}
