% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_qc.R
\name{reads_quality_plot}
\alias{reads_quality_plot}
\title{Reads Quality barplot function}
\usage{
reads_quality_plot(plot_data, output = NULL)
}
\arguments{
\item{plot_data}{reads quality dataframe}

\item{output}{out put prefix, default is NULL, don't output file}
}
\description{
Reads Quality barplot function
}
\examples{
reads_quality_plot(rq_test_data)
}
