% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_themes.R
\name{theme_onmath}
\alias{theme_onmath}
\title{theme for general NGS figures with ggplot}
\usage{
theme_onmath(base_size = 14)
}
\arguments{
\item{base_size}{base font size, default is 14}
}
\description{
theme for general NGS figures with ggplot
}
\examples{
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,colour = factor(gear)))
p + theme_onmath()
}
