% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_go.R
\name{om_goseq}
\alias{om_goseq}
\title{goseq analysis}
\usage{
om_goseq(diff_genes, gene_length_df, go_anno_df, out_prefix = NULL)
}
\arguments{
\item{diff_genes}{diff gene vector}

\item{gene_length_df}{gene length dataframe}

\item{go_anno_df}{go annotation file, seperated with ","}

\item{out_prefix}{out putfile prefix, default is NULL, don't output file}
}
\description{
goseq analysis
}
\examples{
diff_genes <- go_test_data_list[['test_diff_genes']]
gene_length_df <- go_test_data_list[['test_gene_len']]
go_anno_df <- go_test_data_list[['test_go_anno']]
goseq_output <- run_goseq(diff_genes, gene_length_df, go_anno_df)
head(goseq_output, 4)
}
